feat(frontend): improve UI modularity and theme
