+++ frontend/src/theme/ColorModeContext.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+const ColorModeContext = createContext({
+  toggleColorMode: () => {}
+});
+
+export default ColorModeContext;
